requirements.txt
# CableOrca Requirements.
# Instructions:
# "pip install -r requirements.txt"
ipwhois==1.2.0
mac_vendor_lookup==0.1.12
matplotlib==3.7.0
Pillow==9.4.0
requests==2.28.2
scapy==2.5.0
speedtest-cli==2.1.3
